spacetimedb

Attribute Macro table

#[table]

Declares a table with a particular row type.

This attribute is applied to a struct type with named fields. This derives Serialize, Deserialize, SpacetimeType, and Debug for the annotated struct.

Elements of the struct type are NOT automatically inserted into any global table. They are regular structs, with no special behavior. In particular, modifying them does not automatically modify the database!

Instead, a type implementing Table<Row = Self> is generated. This can be looked up in a ReducerContext using ctx.db.{table_name}(). This type represents a handle to a database table, and can be used to iterate and modify the table’s elements. It is a view of the entire table – the entire set of rows at the time of the reducer call.

Example

use spacetimedb::{table, ReducerContext};

#[table(name = user, public,
        index(name = popularity_and_username, btree(columns = [popularity, username])),
)]
pub struct User {
    #[auto_inc]
    #[primary_key]
    pub id: u32,
    #[unique]
    pub username: String,
    #[index(btree)]
    pub popularity: u32,
}

fn demo(ctx: &ReducerContext) {
    // Use the name of the table to get a struct
    // implementing `spacetimedb::Table<Row = User>`.
    let user: user__TableHandle = ctx.db.user();

    // You can use methods from `spacetimedb::Table`
    // on the table.
    log::debug!("User count: {}", user.count());
    for user in user.iter() {
        log::debug!("{:?}", user);
    }

    // For every `#[index(btree)]`, the table has an extra method
    // for getting a corresponding `spacetimedb::BTreeIndex`.
    let by_popularity: RangedIndex<_, (u32,), _> =
        user.popularity();
    for popular_user in by_popularity.filter(95..) {
        log::debug!("Popular user: {:?}", popular_user);
    }

    // There are similar methods for multi-column indexes.
    let by_popularity_and_username: RangedIndex<_, (u32, String), _> = user.popularity_and_username();
    for popular_user in by_popularity.filter((100, "a"..)) {
        log::debug!("Popular user whose name starts with 'a': {:?}", popular_user);
    }

    // For every `#[unique]` or `#[primary_key]` field,
    // the table has an extra method that allows getting a
    // corresponding `spacetimedb::UniqueColumn`.
    let by_username: spacetimedb::UniqueColumn<_, String, _> = user.id();
    by_username.delete(&"test_user".to_string());
}

See Table, RangedIndex, and UniqueColumn for more information on the methods available on these types.

Browsing generated documentation

The #[table] macro generates different APIs depending on the contents of your table.

To browse the complete generated API for your tables, run cargo doc in your SpacetimeDB module project. Navigate to [YOUR PROJECT/target/doc/spacetime_module/index.html in your file explorer, and right click -> open it in a web browser.

For the example above, we would see three items:

A struct User. This is the struct you declared. It stores rows of the table user.

A struct user__TableHandle. This is an opaque handle that allows you to interact with the table user.

A trait user containing a single fn user(&self) -> user__TableHandle. This trait is implemented for the db field of a ReducerContext, allowing you to get a user__TableHandle using ctx.db.user().

Macro arguments

The #[table(...)] attribute accepts any number of the following arguments, separated by commas.

Multiple table annotations can be present on the same type. This will generate multiple tables of the same row type, but with different names.

name

Specify the name of the table in the database. The name can be any valid Rust identifier.

The table name is used to get a handle to the table from a ReducerContext. For a table table, use ctx.db.{table}() to do this. For example:

 #[table(name = user)]
 pub struct User {
     #[auto_inc]
     #[primary_key]
     pub id: u32,
     #[unique]
     pub username: String,
     #[index(btree)]
     pub popularity: u32,
 }
 #[reducer]
 fn demo(ctx: &ReducerContext) {
     let user: user__TableHandle = ctx.db.user();
 }

public and private

Tables are private by default. This means that clients cannot read their contents or see that they exist.

If you’d like to make your table publicly accessible by clients, put public in the macro arguments (e.g. #[spacetimedb::table(public)]). You can also specify private if you’d like to be specific.

This is fully separate from Rust’s module visibility system; pub struct or pub(crate) struct do not affect the table visibility, only the visibility of the items in your own source code.

index(...)

You can specify an index on one or more of the table’s columns with the syntax: index(name = my_index, btree(columns = [a, b, c]))

You can also just put #[index(btree)] on the field itself if you only need a single-column index; see column attributes below.

A table may declare any number of indexes.

You can use indexes to efficiently filter and delete rows. This is encapsulated in the struct RangedIndex.

For a table table and an index index, use:

ctx.db.{table}().{index}()

to get a RangedIndex for a ReducerContext.

For example:

let by_id_and_username: spacetimedb::RangedIndex<_, (u32, String), _> =
    ctx.db.user().by_id_and_username();

scheduled(reducer_name)

Used to declare a scheduled reducer.

The annotated struct type must have at least the following fields:

scheduled_id: u64

scheduled_at: ScheduleAt

Column (field) attributes

#[auto_inc]

Creates an auto-increment constraint.

When a row is inserted with the annotated field set to 0 (zero), the sequence is incremented, and this value is used instead.

Can only be used on numeric types.

May be combined with indexes or unique constraints.

Note that using #[auto_inc] on a field does not also imply #[primary_key] or #[unique]. If those semantics are desired, those attributes should also be used.

When #[auto_inc] is combined with a unique key, be wary not to manually insert values larger than the allocated sequence value. In this case, the sequence will eventually catch up, allocate a value that’s already present, and cause a unique constraint violation.

#[unique]

Creates an unique constraint and index for the annotated field.

You can find, update, and delete rows by their unique columns. This is encapsulated in the struct UniqueColumn.

For a table table and a column column, use:

ctx.db.{table}().{column}()`

to get a UniqueColumn from a ReducerContext.

For example:

let by_username: spacetimedb::UniqueColumn<_, String, _> = ctx.db.user().username();

When there is a unique column constraint on the table, insertion can fail if a uniqueness constraint is violated. If we insert two rows which have the same value of a unique column, the second will fail. This will be via a panic with Table::insert or via a Result::Err with Table::try_insert.

For example:

use spacetimedb::{
    table,
    reducer,
    ReducerContext,
    // Make sure to import the `Table` trait to use `insert` or `try_insert`.
    Table
};

type CountryCode = String;

#[table(name = country)]
struct Country {
    #[unique]
    code: CountryCode,
    national_bird: String
}

#[reducer]
fn insert_unique_demo(ctx: &ReducerContext) {
    let result = ctx.db.country().try_insert(Country {
        code: "AU".into(), national_bird: "Emu".into()
    });
    assert!(result.is_ok());

    let result = ctx.db.country().try_insert(Country {
        code: "AU".into(), national_bird: "Great Egret".into()
        // Whoops, this was Austria's national bird, not Australia's.
        // We should have used the country code "AT", not "AU".
    });
    // since there's already a country in the database with the code "AU",
    // SpacetimeDB gives us an error.
    assert!(result.is_err());

    // The following line would panic, since we use `insert` rather than `try_insert`.
    // let result = ctx.db.country().insert(Country { code: "CN".into(), national_bird: "Blue Magpie".into() });

    // If we wanted to *update* the row for Australia, we can use the `update` method of `UniqueIndex`.
    // The following line will succeed:
    ctx.db.country().code().update(Country {
        code: "AU".into(), national_bird: "Australian Emu".into()
    });
}

#[primary_key]

Implies #[unique]. Also generates additional methods client-side for handling updates to the table.

#[index(btree)]

Creates a single-column index with the specified algorithm.

It is an error to specify this attribute together with #[unique]. Unique constraints implicitly create a unique index, which is accessed using the UniqueColumn struct instead of the RangedIndex struct.

The created index has the same name as the column.

For a table table and an indexed column, use:

ctx.db.{table}().{column}()

to get a RangedIndex from a ReducerContext.

For example:

ctx.db.cities().latitude()

Generated code

For each [table(name = {name})] annotation on a type {T}, generates a struct {name}__TableHandle implementing Table<Row={T}>, and a trait that allows looking up such a {name}Handle in a ReducerContext.

The struct {name}__TableHandle is public and lives next to the row struct. Users are encouraged not to write the name of this table handle struct, or to store table handles in variables; operate through a ReducerContext instead.

For each named index declaration, add a method to {name}__TableHandle for getting a corresponding RangedIndex.

For each field with a #[unique] or #[primary_key] annotation, add a method to {name}Handle for getting a corresponding UniqueColumn.

The following pseudocode illustrates the general idea. Curly braces are used to indicate templated names.

use spacetimedb::{RangedIndex, UniqueColumn, Table, DbView};

// This generated struct is hidden and cannot be directly accessed.
struct {name}__TableHandle { /* ... */ };

// It is a table handle.
impl Table for {name}__TableHandle {
    type Row = {T};
    /* ... */
}

// It can be looked up in a `ReducerContext`,
// using `ctx.db().{name}()`.
trait {name} {
    fn {name}(&self) -> Row = {T}>;
}
impl {name} for <ReducerContext as DbContext>::DbView { /* ... */ }

// Once looked up, it can be used to look up indexes.
impl {name}Handle {
    // For each `#[unique]` or `#[primary_key]` field `{field}` of type `{F}`:
    fn {field}(&self) -> UniqueColumn<_, {F}, _> { /* ... */ };

    // For each named index `{index}` on fields of type `{(F1, ..., FN)}`:
    fn {index}(&self) -> RangedIndex<_, {(F1, ..., FN)}, _>;
}