spacetimedb

Attribute Macro view

#[view]

Marks a function as a spacetimedb view.

A view is a function with read-only access to the database.

The first argument of a view is always a &ViewContext or &AnonymousViewContext. The former can only read from the database whereas latter can also access info about the caller.

After this, a view can take any number of arguments just like reducers. These arguments must implement the SpacetimeType, Serialize, and Deserialize traits. All of these traits can be derived at once by marking a type with #[derive(SpacetimeType)].

Views return Vec<T> or Option<T> where T is a SpacetimeType.

use spacetimedb::{view, table, AnonymousViewContext, SpacetimeType, ViewContext};
use spacetimedb_lib::Identity;

#[table(accessor = player)]
struct Player {
    #[auto_inc]
    #[primary_key]
    id: u64,

    #[unique]
    identity: Identity,

    #[index(btree)]
    level: u32,
}

impl Player {
    fn merge(self, location: Location) -> PlayerAndLocation {
        PlayerAndLocation {
            player_id: self.id,
            level: self.level,
            x: location.x,
            y: location.y,
        }
    }
}

#[derive(SpacetimeType)]
struct PlayerId {
    id: u64,
}

#[derive(SpacetimeType)]
struct PlayerCount {
    count: u64,
}

#[table(accessor = location, index(accessor = coordinates, btree(columns = [x, y])))]
struct Location {
    #[unique]
    player_id: u64,
    x: u64,
    y: u64,
}

#[derive(SpacetimeType)]
struct PlayerAndLocation {
    player_id: u64,
    level: u32,
    x: u64,
    y: u64,
}

// A view that selects at most one row from a table
#[view(accessor = my_player, public)]
fn my_player(ctx: &ViewContext) -> Option<Player> {
    ctx.db.player().identity().find(ctx.sender())
}

// An example of column projection
#[view(accessor = my_player_id, public)]
fn my_player_id(ctx: &ViewContext) -> Option<PlayerId> {
    ctx.db.player().identity().find(ctx.sender()).map(|Player { id, .. }| PlayerId { id })
}

// A view that counts the number of rows in a table
#[view(accessor = player_count, public)]
fn player_count(ctx: &AnonymousViewContext) -> Option<PlayerCount> {
    Some(PlayerCount {
        count: ctx.db.player().count(),
    })
}

// An example that is analogous to a semijoin in sql
#[view(accessor = players_at_coordinates, public)]
fn players_at_coordinates(ctx: &AnonymousViewContext) -> Vec<Player> {
    ctx
        .db
        .location()
        .coordinates()
        .filter((3u64, 5u64))
        .filter_map(|location| ctx.db.player().id().find(location.player_id))
        .collect()
}

// An example of a join that combines fields from two different tables
#[view(accessor = players_with_coordinates, public)]
fn players_with_coordinates(ctx: &AnonymousViewContext) -> Vec<PlayerAndLocation> {
    ctx
        .db
        .location()
        .coordinates()
        .filter((3u64, 5u64))
        .filter_map(|location| ctx
            .db
            .player()
            .id()
            .find(location.player_id)
            .map(|player| player.merge(location))
        )
        .collect()
}

Just like reducers, views are limited in their ability to interact with the outside world. They have no access to any network or filesystem interfaces. Calling methods from std::io, std::net, or std::fs will result in runtime errors.

Views are callable by reducers and other views simply by passing their ViewContext.. This is a regular function call. The callee will run within the caller’s transaction.