spacetimedb

Enum AlgebraicValue

pub enum AlgebraicValue {

    Min,
    Sum(SumValue),
    Product(ProductValue),
    Array(ArrayValue),
    Bool(bool),
    I8(i8),
    U8(u8),
    I16(i16),
    U16(u16),
    I32(i32),
    U32(u32),
    I64(i64),
    U64(u64),
    I128(Packed<i128>),
    U128(Packed<u128>),
    I256(Box<I256>),
    U256(Box<U256>),
    F32(ConstrainedFloat<f32, UnitConstraint<f32>>),
    F64(ConstrainedFloat<f64, UnitConstraint<f64>>),
    String(Box<str>),
    Max,

}

A value in SATS typed at some [AlgebraicType].

Values are type erased, so they do not store their type. This is important mainly for space efficiency, including network latency and bandwidth.

These are only values and not expressions. That is, they are canonical and cannot be simplified further by some evaluation. So forms like 42 + 24 are not represented in an AlgebraicValue.

Variants

Min

The minimum value in the total ordering. Cannot be serialized and only exists to facilitate range index scans. This variant must always be first.

Sum(SumValue)

A structural sum value.

Given a sum type { N_0(T_0), N_1(T_1), ..., N_n(T_n) } where N_i denotes a variant name and where T_i denotes the type the variant stores, a sum value makes a specific choice as to the variant. So for example, we might chose N_1(T_1) and represent this choice with (1, v) where v is a value of type T_1.

Product(ProductValue)

A structural product value.

Given a product type { N_0: T_0, N_1: T_1, ..., N_n: T_n } where N_i denotes a field / element name and where T_i denotes the type the field stores, a product value stores a value v_i of type T_i for each field N_i.

Array(ArrayValue)

A homogeneous array of AlgebraicValues. The array has the type [AlgebraicType::Array(elem_ty)][AlgebraicType::Array].

The contained values are stored packed in a representation appropriate for their type. See [ArrayValue] for details on the representation.

Bool(bool)

A bool value of type [AlgebraicType::Bool].

I8(i8)

An i8 value of type [AlgebraicType::I8].

U8(u8)

A u8 value of type [AlgebraicType::U8].

I16(i16)

An i16 value of type [AlgebraicType::I16].

U16(u16)

A u16 value of type [AlgebraicType::U16].

I32(i32)

An i32 value of type [AlgebraicType::I32].

U32(u32)

A u32 value of type [AlgebraicType::U32].

I64(i64)

An i64 value of type [AlgebraicType::I64].

U64(u64)

A u64 value of type [AlgebraicType::U64].

I128(Packed<i128>)

An i128 value of type [AlgebraicType::I128].

We pack these to shrink AlgebraicValue.

U128(Packed<u128>)

A u128 value of type [AlgebraicType::U128].

We pack these to to shrink AlgebraicValue.

I256(Box<I256>)

An [i256] value of type [AlgebraicType::I256].

We box these up to shrink AlgebraicValue.

U256(Box<U256>)

A [u256] value of type [AlgebraicType::U256].

We pack these to shrink AlgebraicValue.

F32(ConstrainedFloat<f32, UnitConstraint<f32>>)

A totally ordered [F32] value of type [AlgebraicType::F32].

All floating point values defined in IEEE-754 are supported. However, unlike the primitive f32, a total order is established.

F64(ConstrainedFloat<f64, UnitConstraint<f64>>)

A totally ordered [F64] value of type [AlgebraicType::F64].

All floating point values defined in IEEE-754 are supported. However, unlike the primitive f64, a total order is established.

String(Box<str>)

A UTF-8 string value of type [AlgebraicType::String].

Uses Rust’s standard representation of strings.

Max

The maximum value in the total ordering. Cannot be serialized and only exists to facilitate range index scans. This variant must always be last.

Implementations

impl AlgebraicValue

pub fn is_min(&self) -> bool

Returns true if this is a AlgebraicValue::Min, otherwise false

pub fn is_sum(&self) -> bool

Returns true if this is a AlgebraicValue::Sum, otherwise false

pub fn as_sum_mut(&mut self) -> Option<&mut SumValue>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::Sum, otherwise None

pub fn as_sum(&self) -> Option<&SumValue>

Optionally returns references to the inner fields if this is a AlgebraicValue::Sum, otherwise None

pub fn into_sum(self) -> Result<SumValue, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::Sum, otherwise returns back the enum in the Err case of the result

pub fn is_product(&self) -> bool

Returns true if this is a AlgebraicValue::Product, otherwise false

pub fn as_product_mut(&mut self) -> Option<&mut ProductValue>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::Product, otherwise None

pub fn as_product(&self) -> Option<&ProductValue>

Optionally returns references to the inner fields if this is a AlgebraicValue::Product, otherwise None

pub fn into_product(self) -> Result<ProductValue, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::Product, otherwise returns back the enum in the Err case of the result

pub fn is_array(&self) -> bool

Returns true if this is a AlgebraicValue::Array, otherwise false

pub fn as_array_mut(&mut self) -> Option<&mut ArrayValue>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::Array, otherwise None

pub fn as_array(&self) -> Option<&ArrayValue>

Optionally returns references to the inner fields if this is a AlgebraicValue::Array, otherwise None

pub fn into_array(self) -> Result<ArrayValue, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::Array, otherwise returns back the enum in the Err case of the result

pub fn is_bool(&self) -> bool

Returns true if this is a AlgebraicValue::Bool, otherwise false

pub fn as_bool_mut(&mut self) -> Option<&mut bool>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::Bool, otherwise None

pub fn as_bool(&self) -> Option<&bool>

Optionally returns references to the inner fields if this is a AlgebraicValue::Bool, otherwise None

pub fn into_bool(self) -> Result<bool, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::Bool, otherwise returns back the enum in the Err case of the result

pub fn is_i8(&self) -> bool

Returns true if this is a AlgebraicValue::I8, otherwise false

pub fn as_i8_mut(&mut self) -> Option<&mut i8>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I8, otherwise None

pub fn as_i8(&self) -> Option<&i8>

Optionally returns references to the inner fields if this is a AlgebraicValue::I8, otherwise None

pub fn into_i8(self) -> Result<i8, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I8, otherwise returns back the enum in the Err case of the result

pub fn is_u8(&self) -> bool

Returns true if this is a AlgebraicValue::U8, otherwise false

pub fn as_u8_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U8, otherwise None

pub fn as_u8(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a AlgebraicValue::U8, otherwise None

pub fn into_u8(self) -> Result<u8, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U8, otherwise returns back the enum in the Err case of the result

pub fn is_i16(&self) -> bool

Returns true if this is a AlgebraicValue::I16, otherwise false

pub fn as_i16_mut(&mut self) -> Option<&mut i16>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I16, otherwise None

pub fn as_i16(&self) -> Option<&i16>

Optionally returns references to the inner fields if this is a AlgebraicValue::I16, otherwise None

pub fn into_i16(self) -> Result<i16, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I16, otherwise returns back the enum in the Err case of the result

pub fn is_u16(&self) -> bool

Returns true if this is a AlgebraicValue::U16, otherwise false

pub fn as_u16_mut(&mut self) -> Option<&mut u16>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U16, otherwise None

pub fn as_u16(&self) -> Option<&u16>

Optionally returns references to the inner fields if this is a AlgebraicValue::U16, otherwise None

pub fn into_u16(self) -> Result<u16, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U16, otherwise returns back the enum in the Err case of the result

pub fn is_i32(&self) -> bool

Returns true if this is a AlgebraicValue::I32, otherwise false

pub fn as_i32_mut(&mut self) -> Option<&mut i32>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I32, otherwise None

pub fn as_i32(&self) -> Option<&i32>

Optionally returns references to the inner fields if this is a AlgebraicValue::I32, otherwise None

pub fn into_i32(self) -> Result<i32, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I32, otherwise returns back the enum in the Err case of the result

pub fn is_u32(&self) -> bool

Returns true if this is a AlgebraicValue::U32, otherwise false

pub fn as_u32_mut(&mut self) -> Option<&mut u32>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U32, otherwise None

pub fn as_u32(&self) -> Option<&u32>

Optionally returns references to the inner fields if this is a AlgebraicValue::U32, otherwise None

pub fn into_u32(self) -> Result<u32, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U32, otherwise returns back the enum in the Err case of the result

pub fn is_i64(&self) -> bool

Returns true if this is a AlgebraicValue::I64, otherwise false

pub fn as_i64_mut(&mut self) -> Option<&mut i64>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I64, otherwise None

pub fn as_i64(&self) -> Option<&i64>

Optionally returns references to the inner fields if this is a AlgebraicValue::I64, otherwise None

pub fn into_i64(self) -> Result<i64, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I64, otherwise returns back the enum in the Err case of the result

pub fn is_u64(&self) -> bool

Returns true if this is a AlgebraicValue::U64, otherwise false

pub fn as_u64_mut(&mut self) -> Option<&mut u64>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U64, otherwise None

pub fn as_u64(&self) -> Option<&u64>

Optionally returns references to the inner fields if this is a AlgebraicValue::U64, otherwise None

pub fn into_u64(self) -> Result<u64, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U64, otherwise returns back the enum in the Err case of the result

pub fn is_i128(&self) -> bool

Returns true if this is a AlgebraicValue::I128, otherwise false

pub fn as_i128_mut(&mut self) -> Option<&mut Packed<i128>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I128, otherwise None

pub fn as_i128(&self) -> Option<&Packed<i128>>

Optionally returns references to the inner fields if this is a AlgebraicValue::I128, otherwise None

pub fn into_i128(self) -> Result<Packed<i128>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I128, otherwise returns back the enum in the Err case of the result

pub fn is_u128(&self) -> bool

Returns true if this is a AlgebraicValue::U128, otherwise false

pub fn as_u128_mut(&mut self) -> Option<&mut Packed<u128>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U128, otherwise None

pub fn as_u128(&self) -> Option<&Packed<u128>>

Optionally returns references to the inner fields if this is a AlgebraicValue::U128, otherwise None

pub fn into_u128(self) -> Result<Packed<u128>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U128, otherwise returns back the enum in the Err case of the result

pub fn is_i256(&self) -> bool

Returns true if this is a AlgebraicValue::I256, otherwise false

pub fn as_i256_mut(&mut self) -> Option<&mut Box<I256>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::I256, otherwise None

pub fn as_i256(&self) -> Option<&Box<I256>>

Optionally returns references to the inner fields if this is a AlgebraicValue::I256, otherwise None

pub fn into_i256(self) -> Result<Box<I256>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::I256, otherwise returns back the enum in the Err case of the result

pub fn is_u256(&self) -> bool

Returns true if this is a AlgebraicValue::U256, otherwise false

pub fn as_u256_mut(&mut self) -> Option<&mut Box<U256>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::U256, otherwise None

pub fn as_u256(&self) -> Option<&Box<U256>>

Optionally returns references to the inner fields if this is a AlgebraicValue::U256, otherwise None

pub fn into_u256(self) -> Result<Box<U256>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::U256, otherwise returns back the enum in the Err case of the result

pub fn is_f32(&self) -> bool

Returns true if this is a AlgebraicValue::F32, otherwise false

pub fn as_f32_mut( &mut self, ) -> Option<&mut ConstrainedFloat<f32, UnitConstraint<f32>>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::F32, otherwise None

pub fn as_f32(&self) -> Option<&ConstrainedFloat<f32, UnitConstraint<f32>>>

Optionally returns references to the inner fields if this is a AlgebraicValue::F32, otherwise None

pub fn into_f32( self, ) -> Result<ConstrainedFloat<f32, UnitConstraint<f32>>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::F32, otherwise returns back the enum in the Err case of the result

pub fn is_f64(&self) -> bool

Returns true if this is a AlgebraicValue::F64, otherwise false

pub fn as_f64_mut( &mut self, ) -> Option<&mut ConstrainedFloat<f64, UnitConstraint<f64>>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::F64, otherwise None

pub fn as_f64(&self) -> Option<&ConstrainedFloat<f64, UnitConstraint<f64>>>

Optionally returns references to the inner fields if this is a AlgebraicValue::F64, otherwise None

pub fn into_f64( self, ) -> Result<ConstrainedFloat<f64, UnitConstraint<f64>>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::F64, otherwise returns back the enum in the Err case of the result

pub fn is_string(&self) -> bool

Returns true if this is a AlgebraicValue::String, otherwise false

pub fn as_string_mut(&mut self) -> Option<&mut Box<str>>

Optionally returns mutable references to the inner fields if this is a AlgebraicValue::String, otherwise None

pub fn as_string(&self) -> Option<&Box<str>>

Optionally returns references to the inner fields if this is a AlgebraicValue::String, otherwise None

pub fn into_string(self) -> Result<Box<str>, AlgebraicValue>

Returns the inner fields if this is a AlgebraicValue::String, otherwise returns back the enum in the Err case of the result

pub fn is_max(&self) -> bool

Returns true if this is a AlgebraicValue::Max, otherwise false

impl AlgebraicValue

pub fn take(&mut self) -> AlgebraicValue

Extract the value and replace it with a dummy one that is cheap to make.

pub fn as_bytes(&self) -> Option<&[u8]>

Interpret the value as a byte slice or None if it isn’t a byte slice.

pub fn unit() -> AlgebraicValue

The canonical unit value defined as the nullary product value ().

The type of UNIT is ().

pub const fn Bytes(v: Box<[u8]>) -> AlgebraicValue

Returns an AlgebraicValue representing v: Box<[u8]>.

pub fn into_bytes(self) -> Result<Box<[u8]>, AlgebraicValue>

Converts self into a byte string, if applicable.

pub fn into_option(self) -> Result<Option<AlgebraicValue>, AlgebraicValue>

Converts self into an Option<AlgebraicValue>, if applicable.

pub fn OptionSome(v: AlgebraicValue) -> AlgebraicValue

Returns an AlgebraicValue for some: v.

The some variant is assigned the tag 0.

pub fn OptionNone() -> AlgebraicValue

Returns an AlgebraicValue for none.

The none variant is assigned the tag 1.

pub fn into_result( self, ) -> Result<Result<AlgebraicValue, AlgebraicValue>, AlgebraicValue>

Converts self into Result<Result<AlgebraicValue, AlgebraicValue>, Self>, if applicable.

pub fn ResultOk(v: AlgebraicValue) -> AlgebraicValue

Returns an AlgebraicValue for Ok: v.

The Ok variant is assigned the tag 0.

pub fn ResultErr(v: AlgebraicValue) -> AlgebraicValue

Returns an AlgebraicValue for Err: v.

The Err variant is assigned the tag 1.

pub fn sum(tag: u8, value: AlgebraicValue) -> AlgebraicValue

Returns an AlgebraicValue representing a sum value with tag and value.

pub fn enum_simple(tag: u8) -> AlgebraicValue

Returns an AlgebraicValue representing a sum value with tag and empty AlgebraicValue::product, that is valid for simple enums without payload.

pub fn product(elements: impl Into<ProductValue>) -> AlgebraicValue

Returns an AlgebraicValue representing a product value with the given elements.

pub fn type_of(&self) -> Option<AlgebraicType>

Infer the [AlgebraicType] of an AlgebraicValue.

This function is partial as type inference is not possible for AlgebraicValue in the case of sums. Thus the method only answers for the decidable subset.

A note on sums

The type of a sum value must be a sum type and not a product type. Suppose x.tag is for the variant VarName(VarType). Then VarType is not the same type as { VarName(VarType) | r } where r represents a polymorphic variants component.

To assign this a correct type we either have to store the type with the value r alternatively, we must have polymorphic variants (see row polymorphism) and derive the correct variant name.

pub fn is_numeric_zero(&self) -> bool

Returns whether this value represents a numeric zero.

Can only be true where the type is numeric.

pub fn from_i128(ty: &AlgebraicType, value: i128) -> Option<AlgebraicValue>

Constructs an AlgebraicValue from an i128 according to the given AlgebraicType.

Returns None if the type is not a supported integer type.

impl AlgebraicValue

pub fn decode<'a>( ty: &<AlgebraicValue as Value>::Type, bytes: &mut impl BufReader<'a>, ) -> Result<AlgebraicValue, DecodeError>

Decode a value from bytes typed at ty.

pub fn decode_smallvec<'a>( ty: &<AlgebraicValue as Value>::Type, bytes: &mut impl BufReader<'a>, ) -> Result<SmallVec<[AlgebraicValue; 1]>, DecodeError>

Decode a vector of values from bytes with each value typed at ty.

pub fn encode(&self, bytes: &mut impl BufWriter)