spacetimedb

Struct Errno

pub struct Errno(/* private fields */);

Error values used in the safe bindings API.

Implementations

impl Errno

pub const HOST_CALL_FAILURE: Errno

ABI called by host returned an error

pub const NOT_IN_TRANSACTION: Errno

ABI call can only be made while in a transaction

pub const BSATN_DECODE_ERROR: Errno

Couldn’t decode the BSATN to the expected type

pub const NO_SUCH_TABLE: Errno

No such table

pub const NO_SUCH_INDEX: Errno

No such index

pub const NO_SUCH_ITER: Errno

The provided row iterator is not valid

pub const NO_SUCH_CONSOLE_TIMER: Errno

The provided console timer does not exist

pub const NO_SUCH_BYTES: Errno

The provided bytes source or sink is not valid

pub const NO_SPACE: Errno

The provided sink has no more space left

pub const BUFFER_TOO_SMALL: Errno

The provided buffer is not large enough to store the data

pub const UNIQUE_ALREADY_EXISTS: Errno

Value with given unique identifier already exists

pub const SCHEDULE_AT_DELAY_TOO_LONG: Errno

Specified delay in scheduling row was too long

pub const INDEX_NOT_UNIQUE: Errno

The index was not unique

pub const NO_SUCH_ROW: Errno

The row was not found, e.g., in an update call

impl Errno

pub const fn message(self) -> Option<&'static str>

Returns a description of the errno value, if any.

pub const fn from_code(code: u16) -> Option<Errno>

Converts the given code to an error number in Errno’s representation.

pub const fn code(self) -> u16

Converts this errno into a primitive error code.