spacetimedb

Struct Identity

pub struct Identity { /* private fields */ }

An Identity for something interacting with the database.

An Identity is a 256-bit unsigned integer. These are encoded in various ways.

In JSON, an Identity is represented as a hexadecimal number wrapped in a string, "0x[64 hex characters]".

In BSATN, an Identity is represented as a LITTLE-ENDIAN number 32 bytes long.

In memory, an Identity is stored as a 256-bit number with the endianness of the host system.

If you are manually converting a hexadecimal string to a byte array like so:

"0xb0b1b2..."
->
[0xb0, 0xb1, 0xb2, ...]

Make sure you call Identity::from_be_byte_array and NOT Identity::from_byte_array. The standard way of writing hexadecimal numbers follows a big-endian convention, if you index the characters in written text in increasing order from left to right.

Implementations

impl Identity

pub const ZERO: Identity

The 0x0 Identity

pub const ONE: Identity

The 0x1 Identity

pub const fn from_byte_array(bytes: [u8; 32]) -> Identity

Create an Identity from a LITTLE-ENDIAN byte array.

If you are parsing an Identity from a string, you probably want from_be_byte_array instead.

pub const fn from_be_byte_array(bytes: [u8; 32]) -> Identity

Create an Identity from a BIG-ENDIAN byte array.

This method is the correct choice if you have converted the bytes of a hexadecimal-formatted Identity to a byte array in the following way:

"0xb0b1b2..."
->
[0xb0, 0xb1, 0xb2, ...]

pub const fn from_u256(__identity__: U256) -> Identity

Converts __identity__: u256 to Identity.

pub const fn to_u256(&self) -> U256

Converts this identity to an u256.

pub fn from_claims(issuer: &str, subject: &str) -> Identity

pub fn to_byte_array(&self) -> [u8; 32]

Returns this Identity as a byte array.

pub fn to_be_byte_array(&self) -> [u8; 32]

Convert this Identity to a BIG-ENDIAN byte array.

pub fn to_hex(&self) -> HexString<32>

Convert this Identity to a hexadecimal string.

pub fn abbreviate(&self) -> [u8; 8]

Extract the first 8 bytes of this Identity as if it was stored in BIG-ENDIAN format. (That is, the most significant bytes.)

pub fn to_abbreviated_hex(&self) -> HexString<8>

Extract the first 16 characters of this Identity’s hexadecimal representation.

pub fn from_hex(hex: impl AsRef<[u8]>) -> Result<Identity, FromHexError>