spacetimedb

Struct UniqueColumn

pub struct UniqueColumn<Tbl, ColType, Col> { /* private fields */ }

A handle to a unique index on a column. Available for #[unique] and #[primary_key] columns.

For a table table with a column column, use ctx.db.{table}().{column}() to get a UniqueColumn from a ReducerContext.

Example:

use spacetimedb::{table, UniqueColumn, ReducerContext, DbContext};

#[table(name = user)]
struct User {
    #[primary_key]
    id: u32,
    #[unique]
    username: String,
    dog_count: u64
}

fn demo(ctx: &ReducerContext) {
    let user = ctx.db().user();

    let by_id: UniqueColumn<_, u32, _> = user.id();

    let mut example_user: User = by_id.find(357).unwrap();
    example_user.dog_count += 5;
    by_id.update(example_user);

    let by_username: UniqueColumn<_, String, _> = user.username();
    by_username.delete(&"Evil Bob".to_string());
}

Implementations

impl<Tbl: Table, Col: Index + Column<Table = Tbl>> UniqueColumn<Tbl, Col::ColType, Col>

pub fn find(&self, col_val: impl Borrow<Col::ColType>) -> Option<Tbl::Row>
where for<'a> &'a Col::ColType: FilterableValue,

Finds and returns the row where the value in the unique column matches the supplied col_val, or None if no such row is present in the database state.

pub fn delete(&self, col_val: impl Borrow<Col::ColType>) -> bool

Deletes the row where the value in the unique column matches the supplied col_val, if any such row is present in the database state.

Returns true if a row with the specified col_val was previously present and has been deleted, or false if no such row was present.

pub fn update(&self, new_row: Tbl::Row) -> Tbl::Row

Deletes the row where the value in the unique column matches that in the corresponding field of new_row, and then inserts the new_row.

Returns the new row as actually inserted, with computed values substituted for any auto-inc placeholders.

Panics

Panics if no row was previously present with the matching value in the unique column, or if either the delete or the insertion would violate a constraint.