spacetimedb

Attribute Macro reducer

#[reducer]

Marks a function as a spacetimedb reducer.

A reducer is a function with read/write access to the database that can be invoked remotely by clients.

Each reducer call runs in its own database transaction, and its updates to the database are only committed if the reducer returns successfully.

The first argument of a reducer is always a &ReducerContext. This context object allows accessing the database and viewing information about the caller, among other things.

After this, a reducer can take any number of arguments. These arguments must implement the SpacetimeType, Serialize, and Deserialize traits. All of these traits can be derived at once by marking a type with #[derive(SpacetimeType)].

Reducers may return either () or Result<(), E> where E: std::fmt::Display.

use spacetimedb::{reducer, SpacetimeType, ReducerContext};
use log::info;
use std::fmt;

#[reducer]
pub fn hello_world(context: &ReducerContext) {
    info!("Hello, World!");
}

#[reducer]
pub fn add_person(context: &ReducerContext, name: String, age: u16) {
    // add a "person" to the database.
}

#[derive(SpacetimeType, Debug)]
struct Coordinates {
    x: f32,
    y: f32,
}

enum AddPlaceError {
    InvalidCoordinates(Coordinates),
    InvalidName(String),
}

impl fmt::Display for AddPlaceError {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
        match self {
            AddPlaceError::InvalidCoordinates(coords) => {
                write!(f, "invalid coordinates: {coords:?}")
            },
            AddPlaceError::InvalidName(name) => {
                write!(f, "invalid name: {name:?}")
            },
        }
    }
}

#[reducer]
pub fn add_place(
    context: &ReducerContext,
    name: String,
    x: f32,
    y: f32,
    area: f32,
) -> Result<(), AddPlaceError> {
    // ... add a place to the database...
}

Reducers may fail by returning a Result::Err or by panicking. Failures will abort the active database transaction. Any changes to the database made by the failed reducer call will be rolled back.

Reducers are limited in their ability to interact with the outside world. They do not directly return data aside from errors, and have no access to any network or filesystem interfaces. Calling methods from std::io, std::net, or std::fs inside a reducer will result in runtime errors.

Reducers can communicate information to the outside world in two ways:

They can modify tables in the database. See the #[table](#table) macro documentation for information on how to declare and use tables.

They can call logging macros from the log crate. This writes to a private debug log attached to the database. Run spacetime logs <DATABASE_IDENTITY> to browse these.

Reducers are permitted to call other reducers, simply by passing their ReducerContext as the first argument. This is a regular function call, and does not involve any network communication. The callee will run within the caller’s transaction, and any changes made by the callee will be committed or rolled back with the caller.

Lifecycle Reducers

You can specify special lifecycle reducers that are run at set points in the module’s lifecycle. You can have one of each per module.

These reducers cannot be called manually and may not have any parameters except for ReducerContext.

The init reducer

This reducer is marked with #[spacetimedb::reducer(init)]. It is run the first time a module is published and any time the database is cleared. (It does not have to be named init.)

If an error occurs when initializing, the module will not be published.

This reducer can be used to configure any static data tables used by your module. It can also be used to start running scheduled reducers.

The client_connected reducer

This reducer is marked with #[spacetimedb::reducer(client_connected)]. It is run when a client connects to the SpacetimeDB module. Their identity can be found in the sender value of the ReducerContext.

If an error occurs in the reducer, the client will be disconnected.

The client_disconnected reducer

This reducer is marked with #[spacetimedb::reducer(client_disconnected)]. It is run when a client disconnects from the SpacetimeDB module. Their identity can be found in the sender value of the ReducerContext.

If an error occurs in the disconnect reducer, the client is still recorded as disconnected.

Scheduled reducers

In addition to life cycle annotations, reducers can be made scheduled. This allows calling the reducers at a particular time, or in a loop. This can be used for game loops.

The scheduling information for a reducer is stored in a table. This table has two mandatory fields:

A primary key that identifies scheduled reducer calls.

A ScheduleAt field that says when to call the reducer.

Managing timers with a scheduled table is as simple as inserting or deleting rows from the table. This makes scheduling transactional in SpacetimeDB. If a reducer A first schedules B but then errors for some other reason, B will not be scheduled to run.

A ScheduleAt can be created from a spacetimedb::Timestamp, in which case the reducer will be scheduled once, or from a std::time::Duration, in which case the reducer will be scheduled in a loop. In either case the conversion can be performed using Into::into.

use spacetimedb::{table, reducer, ReducerContext, Timestamp, TimeDuration, ScheduleAt, Table};
use log::debug;

// First, we declare the table with scheduling information.

#[table(name = send_message_schedule, scheduled(send_message))]
struct SendMessageSchedule {
    // Mandatory fields:
    // ============================

    /// An identifier for the scheduled reducer call.
    #[primary_key]
    #[auto_inc]
    scheduled_id: u64,

    /// Information about when the reducer should be called.
    scheduled_at: ScheduleAt,

    // In addition to the mandatory fields, any number of fields can be added.
    // These can be used to provide extra information to the scheduled reducer.

    // Custom fields:
    // ============================

    /// The text of the scheduled message to send.
    text: String,
}

// Then, we declare the scheduled reducer.
// The first argument of the reducer should be, as always, a `&ReducerContext`.
// The second argument should be a row of the scheduling information table.

#[reducer]
fn send_message(ctx: &ReducerContext, arg: SendMessageSchedule) -> Result<(), String> {
    let message_to_send = arg.text;

    // ... send the message ...

    Ok(())
}

// Now, we want to actually start scheduling reducers.
// It's convenient to do this inside the `init` reducer.
#[reducer(init)]
fn init(ctx: &ReducerContext) {

    let current_time = ctx.timestamp;

    let ten_seconds = TimeDuration::from_micros(10_000_000);

    let future_timestamp: Timestamp = ctx.timestamp + ten_seconds;
    ctx.db.send_message_schedule().insert(SendMessageSchedule {
        scheduled_id: 1,
        text:"I'm a bot sending a message one time".to_string(),

        // Creating a `ScheduleAt` from a `Timestamp` results in the reducer
        // being called once, at exactly the time `future_timestamp`.
        scheduled_at: future_timestamp.into()
    });

    let loop_duration: TimeDuration = ten_seconds;
    ctx.db.send_message_schedule().insert(SendMessageSchedule {
        scheduled_id: 0,
        text:"I'm a bot sending a message every 10 seconds".to_string(),

        // Creating a `ScheduleAt` from a `Duration` results in the reducer
        // being called in a loop, once every `loop_duration`.
        scheduled_at: loop_duration.into()
    });
}

Scheduled reducers are called on a best-effort basis and may be slightly delayed in their execution when a database is under heavy load.

Restricting scheduled reducers

Scheduled reducers are normal reducers, and may still be called by clients. If a scheduled reducer should only be called by the scheduler, consider beginning it with a check that the caller Identity is the module:

use spacetimedb::{reducer, ReducerContext};

#[reducer]
fn scheduled(ctx: &ReducerContext, args: ScheduledArgs) -> Result<(), String> {
    if ctx.sender != ctx.identity() {
        return Err("Reducer `scheduled` may not be invoked by clients, only via scheduling.".into());
    }
    // Reducer body...
}