spacetimedb

Struct StdbRng

pub struct StdbRng { /* private fields */ }

A reference to the random number generator for this reducer call.

An instance can be obtained via ReducerContext::rng(). Import rand::Rng in order to access many useful random algorithms.

StdbRng uses the same PRNG as rand’s StdRng. Note, however, that because it is seeded from a publically-known timestamp, it is not cryptographically secure.

You may be looking for a level of reproducibility that’s finer-grained than “if it happens at the exact same time, you get the same result” – if so, just seed an StdRng directly, or use another rng like those listed here. Just note that you must not store any state, including an rng, in a global variable or any other in-WASM side channel. Any and all state persisted across reducer calls must be stored in the database.