spacetimedb

Trait DbContext

pub trait DbContext {
    type DbView;

    // Required method
    fn db(&self) -> &Self::DbView;
}

A handle on a database with a particular table schema.

Required Associated Types

type DbView

A view into the tables of a database.

This type is specialized on the database’s particular schema.

Methods on the DbView type will allow querying tables defined by the module.

Required Methods

fn db(&self) -> &Self::DbView

Get a view into the tables.

This method is provided for times when a programmer wants to be generic over the DbContext type. Concrete-typed code is expected to read the .db field off the particular DbContext implementor. Currently, being this generic is only meaningful in clients, as ReducerContext is the only implementor of DbContext within modules.