spacetimedb

Trait FilterableValue

pub trait FilterableValue: Serialize + Private {
    type Column;
}

Types which can appear as an argument to an index filtering operation for a column of type Column.

Types which can appear specifically as a terminating bound in a BTree index, which may be a range, instead use IndexScanRangeBoundsTerminator. Because SpacetimeDB supports a only restricted set of types as index keys, only a small set of Column types have corresponding FilterableValue implementations. Specifically, these types are:

Signed and unsigned integers of various widths.

bool.

String, which is also filterable with &str.

Identity.

ConnectionId.

Hash.

No-payload enums annotated with #[derive(SpacetimeType)]. No-payload enums are sometimes called “plain,” “simple” or “C-style.” They are enums where no variant has any payload data.

Required Associated Types

type Column

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.